Break out WelcomeMsg component, use in portfolio
