+++ client-app/src/desktop/App.js
@@ -1,10 +1,11 @@
-import {br, div, img} from '@xh/hoist/cmp/layout';
+import {img} from '@xh/hoist/cmp/layout';
 import {tabContainer} from '@xh/hoist/cmp/tab';
 import {webSocketIndicator} from '@xh/hoist/cmp/websocket';
-import {hoistCmp, uses, XH} from '@xh/hoist/core';
+import {hoistCmp, uses} from '@xh/hoist/core';
 import {appBar, appBarSeparator} from '@xh/hoist/desktop/cmp/appbar';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {tabSwitcher} from '@xh/hoist/desktop/cmp/tab';
+import {welcomeMsg} from '../core/cmp/WelcomeMsg';
 import xhLogo from '../core/img/xh-toolbox-logo.png';
 import '../core/Toolbox.scss';
 import './App.scss';
@@ -15,8 +16,6 @@ export const App = hoistCmp({
     model: uses(AppModel),
 
     render({model}) {
-        const user = XH.getUser();
-
         return panel({
             tbar: appBar({
                 icon: img({src: xhLogo, onClick: () => model.goHome()}),
@@ -32,17 +31,7 @@ export const App = hoistCmp({
                 appMenuButtonProps: {
                     hideLogoutItem: false,
                     extraItems: [
-                        div({
-                            className: 'tb-welcome-message',
-                            items: [
-                                div({
-                                    className: 'tb-welcome-message__profile-pic',
-                                    item: img({src: user.profilePicUrl}),
-                                    omit: !user.profilePicUrl
-                                }),
-                                `Welcome,`, br(), user.displayName
-                            ]
-                        })
+                        welcomeMsg({multiline: true})
                     ]
                 }
             }),
