+++ client-app/src/desktop/App.scss
@@ -7,32 +7,6 @@
   }
 }
 
-.tb-welcome-message {
-  display: flex;
-  align-items: center;
-  padding: 10px 5px;
-
-  &__profile-pic {
-    flex: none;
-    position: relative;
-    display: inline-block;
-    height: 35px;
-    width: 35px;
-    overflow: hidden;
-    border-radius: 50%;
-    margin-right: 10px;
-
-    img {
-      position: absolute;
-      left: 50%;
-      top: 50%;
-      height: auto;
-      width: 100%;
-      transform: translate(-50%, -50%);
-    }
-  }
-}
-
 .toolbox-tab > .xh-tab-switcher--left > .bp3-tab-list {
   width: 140px;
   border-right: var(--xh-border-solid);
