+++ client-app/src/desktop/common/Wrapper.js
@@ -4,7 +4,7 @@ import {box, table, tbody, td, th, tr} from '@xh/hoist/cmp/layout';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
 import {dockContainer, DockContainerModel} from '@xh/hoist/cmp/dock';
-import {toolboxLink} from '../../core/ToolboxLink';
+import {toolboxLink} from '../../core/cmp/ToolboxLink';
 import './Wrapper.scss';
 
 /**
