+++ client-app/src/examples/portfolio/App.js
@@ -3,6 +3,7 @@ import {hoistCmp, uses} from '@xh/hoist/core';
 import {appBar, appBarSeparator} from '@xh/hoist/desktop/cmp/appbar';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
 import {Icon} from '@xh/hoist/icon';
+import {welcomeMsg} from '../../core/cmp/WelcomeMsg';
 import {AppModel} from './AppModel';
 import {portfolioPanel} from './PortfolioPanel';
 import '../../core/Toolbox.scss';
@@ -21,7 +22,9 @@ export const App = hoistCmp.factory({
                 },
                 appMenuButtonPosition: 'left',
                 rightItems: [
-                    webSocketIndicator({iconOnly: true, marginRight: 4}),
+                    welcomeMsg(),
+                    appBarSeparator(),
+                    webSocketIndicator(),
                     appBarSeparator()
                 ]
             }),
